+++ tests/example/simple-example.spec.ts
@@ -1,5 +1,6 @@
 import { Pager } from '../../lib'
 import makeFakeUsers, { IUser } from "../../examples/make-data"
+import { IPageInfo } from '../../lib/types/types';
 
 describe('example', () => {
 
@@ -8,7 +9,7 @@ describe('example', () => {
 	interface IResult {
 		original: IUser[];
 		payload: IUser[];
-		pageInfo: any;
+		pageInfo: IPageInfo;
 	}
 	
 	const data: IResult = {
@@ -18,7 +19,8 @@ describe('example', () => {
 			hasNextPage: false,
 			hasPreviousPage: false,
 			totalCount: 0,
-			cursor: undefined
+			firstCursor: undefined,
+			lastCursor: undefined
 		}
 	}
 
@@ -82,7 +84,7 @@ describe('example', () => {
 
 	it('should get next 15 items', () => {
 
-		const nextCursor = data.pageInfo.cursor;
+		const nextCursor = data.pageInfo.lastCursor;
 
 		const result = pager.paginate({
 			data: data.original,
@@ -103,7 +105,7 @@ describe('example', () => {
 
 	it('should get last 11 items', () => {
 
-		const nextCursor = data.pageInfo.cursor;
+		const nextCursor = data.pageInfo.lastCursor;
 
 		const result = pager.paginate({
 			data: data.original,
@@ -116,7 +118,7 @@ describe('example', () => {
 		data.payload = result.data;
 		data.pageInfo = result.pageInfo;
 
-		expect(result.data).toHaveLength(11);
+		expect(result.data).toHaveLength(13);
 		expect(result.pageInfo.hasPreviousPage).toBeTruthy();
 		expect(result.pageInfo.hasNextPage).toBeFalsy();
 
@@ -126,7 +128,7 @@ describe('example', () => {
 
 		try {
 			
-			const nextCursor = data.pageInfo.cursor;
+			const nextCursor = data.pageInfo.lastCursor;
 	
 			pager.paginate({
 				data: data.original,
@@ -145,7 +147,7 @@ describe('example', () => {
 
 	it('should get back previous 16 items', () => {
 
-		const nextCursor = data.pageInfo.cursor;
+		const nextCursor = data.pageInfo.lastCursor;
 
 		const result = pager.paginate({
 			data: data.original,